Think About The NavBar
Why? I need to come up a design that looks nice. It
also serve as a guideline to change the code to
address more SoC and DRY principles now that
experimental testing/learning phase is over.
I've decided to Logo on the left, a show new tweets,
and a write new tweets on the right.