Compose Feature
Added a guest writer on the compose Twidder button.
User Tweets will automatically be updated to time line. It will auto refresh
and scroll you directly there. A possible fature would be to grab the timeline
offset so that you refresh directly to your recent tweet at the top of
the page. That will require a rework of fooLooper. Note to self. Create
better helper functions that can withstand outside requiremental changes.
Created a failures.md full of graveyard approaches.