Time Delay: Bug Fixed
Adding a buffer to allow the timers to breathe. This is why tweet.user
was going through after throwing the error. It had to wait for the delayed
function to pass thru the necessary data. If i only operate one before
the latest and TELL the USER, that there are no more tweets when there's
in actuality a cache of 1 or 2 tweets, the system doesn't go haywire.